Refactor: Move delegates into module
