+++ modules/delegates/components/DelegateParticipationMetrics.tsx
@@ -1,0 +1,10 @@
+import { Box } from 'theme-ui';
+import { Delegate } from '../types';
+
+export function DelegateParticipationMetrics({ delegate }: { delegate: Delegate }): React.ReactElement {
+  return (
+    <Box>
+      Vote History
+    </Box>
+  );
+}
