+++ modules/delegates/types/delegate.d.ts
