+++ modules/delegates/types/index.ts
@@ -1,0 +1,2 @@
+export * from './delegate';
+export * from './delegatesAPI';
