+++ pages/delegates/index.tsx
@@ -5,16 +5,16 @@ import { GetStaticProps } from 'next';
 import Head from 'next/head';
 import ErrorPage from 'next/error';
 import { isDefaultNetwork } from 'lib/maker';
-import { fetchDelegates } from 'lib/delegates/fetchDelegates';
-import { DelegateStatusEnum } from 'lib/delegates/constants';
+import { fetchDelegates } from 'modules/delegates/api/fetchDelegates';
+import { DelegateStatusEnum } from 'modules/delegates/delegates.constants';
 import { shuffleArray } from 'lib/common/shuffleArray';
-import { Delegate } from 'types/delegate';
+import { Delegate, DelegatesAPIResponse, DelegatesAPIStats } from 'modules/delegates/types';
 import PrimaryLayout from 'components/layouts/Primary';
 import SidebarLayout from 'components/layouts/Sidebar';
 import Stack from 'components/layouts/Stack';
 import SystemStatsSidebar from 'components/SystemStatsSidebar';
 import ResourceBox from 'components/ResourceBox';
-import { DelegateCard } from 'components/delegations';
+import { DelegateCard } from 'modules/delegates/components';
 import PageLoadingPlaceholder from 'components/PageLoadingPlaceholder';
 import { getNetwork } from 'lib/maker';
 import { fetchJson } from 'lib/utils';
@@ -22,8 +22,7 @@ import { useAnalytics } from 'lib/client/analytics/useAnalytics';
 import { ANALYTICS_PAGES } from 'lib/client/analytics/analytics.constants';
 import useAccountsStore from 'stores/accounts';
 import Link from 'next/link';
-import { DelegatesSystemInfo } from 'components/delegations/DelegatesSystemInfo';
-import { DelegatesAPIResponse, DelegatesAPIStats } from 'types/delegatesAPI';
+import { DelegatesSystemInfo } from 'modules/delegates/components/DelegatesSystemInfo';
 import { getExecutiveProposals } from 'modules/executives/api/fetchExecutives';
 import { CMSProposal } from 'types/proposal';
